requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
python-dotenv
pymongo
certifi
pymongo
pyaml
dill
mlflow
fastapi
uvicorn

#-e .